Move Pipeline Build Decorator to topology2
