+++ frontend/packages/dev-console/src/components/topology2/components/nodes/build-decorators/BuildDecorator.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
 import { Tooltip, TooltipPosition } from '@patternfly/react-core';
-import { WorkloadData } from '../../topology-types';
-import Decorator from '../Decorator';
+import { WorkloadData } from '../../../../topology/topology-types';
+import Decorator from '../../../../topology/shapes/Decorator';
 import { getBuildDecoratorParts } from './build-decorator-utils';
 
 export interface BuildDecoratorProps {
