+++ frontend/packages/dev-console/src/components/topology2/components/nodes/build-decorators/PipelineBuildDecoratorTooltip.scss
