+++ frontend/packages/dev-console/src/components/topology2/components/nodes/build-decorators/build-decorator-utils.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { Status } from '@console/shared';
 import { resourcePathFromModel } from '@console/internal/components/utils';
 import { BuildModel } from '@console/internal/models';
-import { PipelineRunModel } from '../../../../models';
-import { constructCurrentPipeline } from '../../../../utils/pipeline-utils';
-import { WorkloadData } from '../../topology-types';
+import { PipelineRunModel } from '../../../../../models';
+import { constructCurrentPipeline } from '../../../../../utils/pipeline-utils';
+import { WorkloadData } from '../../../../topology/topology-types';
 import PipelineBuildDecoratorTooltip from './PipelineBuildDecoratorTooltip';
 
 type BuildDecoratorData = {
